Implement CurrentUser service

+++ SleekChat/Api/Controllers/UsersController.cs
@@ -15,18 +15,23 @@ namespace SleekChat.Api.Controllers
     public class UsersController : ControllerBase
     {
         private readonly IUserData userData;
+        private readonly ICurrentUser currentUser;
         private readonly ValidationHelper validator;
         private readonly FormatHelper formatter;
+        public readonly HttpHelper httpHelper;
         private KeyValuePair<bool, string> validationResult;
         private readonly IOptions<AppSettings> config;
 
-        public UsersController(IUserData userData, IOptions<AppSettings> config)
+        public UsersController(IUserData userData, IOptions<AppSettings> config, ICurrentUser currentUser)
         {
             this.userData = userData;
             this.config = config;
+            this.currentUser = currentUser;
             validator = new ValidationHelper();
             formatter = new FormatHelper();
+            httpHelper = new HttpHelper();
         }
+
 
         // GET: api/users
         [HttpGet]
@@ -35,17 +40,18 @@ namespace SleekChat.Api.Controllers
             return Ok(formatter.Render(userData.GetAllUsers(), "Users", Operation.Retrieved));
         }
 
+
         // GET: api/users/id
         [HttpGet("{id}")]
         public ActionResult Get(string id)
         {
             validationResult = validator.IsBlank("User Id", id);
-            if (validationResult.Key == false) 
-                return BadRequest(formatter.Render(validationResult)); 
+            if (validationResult.Key == false)
+                return BadRequest(formatter.Render(validationResult));
 
             validationResult = validator.IsValidGuid("User Id", id);
             if (validationResult.Key == false)
-                return BadRequest(formatter.Render(validationResult)); 
+                return BadRequest(formatter.Render(validationResult));
 
             User user = userData.GetUserById(Guid.Parse(id));
             if (user == null)
@@ -53,6 +59,7 @@ namespace SleekChat.Api.Controllers
 
             return Ok(formatter.Render(user, "User", Operation.Retrieved));
         }
+
 
         // POST: api/users/register
         [AllowAnonymous]
@@ -65,7 +72,7 @@ namespace SleekChat.Api.Controllers
             if (validationResult.Key == false)
                 return BadRequest(formatter.Render(validationResult));
 
-            validationResult = validator.IsBlank("email", email); 
+            validationResult = validator.IsBlank("email", email);
             if (validationResult.Key == false)
                 return BadRequest(formatter.Render(validationResult));
 
@@ -73,9 +80,9 @@ namespace SleekChat.Api.Controllers
             if (validationResult.Key == false)
                 return BadRequest(formatter.Render(validationResult));
 
-            validationResult = validator.PasswordsMatch(password, cPassword); 
+            validationResult = validator.PasswordsMatch(password, cPassword);
             if (validationResult.Key == false)
-                return BadRequest(formatter.Render(validationResult)); 
+                return BadRequest(formatter.Render(validationResult));
 
             if (userData.UsernameAlreadyTaken(username, out _))
                 return Conflict(formatter.Render(validator.Result($"The username '{username}' is already taken.")));
@@ -113,7 +120,7 @@ namespace SleekChat.Api.Controllers
         [HttpPut("{id}")]
         public ActionResult Put([FromRoute] string id, [FromBody] RequestBody reqBody)
         {
-            (string username, string email, string password, string cPassword, _) = reqBody;
+            (string username, string email, string password, string cPassword, _) = reqBody;           
 
             validationResult = validator.IsBlank("User Id", id);
             if (validationResult.Key == false)
@@ -126,6 +133,13 @@ namespace SleekChat.Api.Controllers
             User user = userData.GetUserById(Guid.Parse(id));
             if (user == null)
                 return NotFound(formatter.Render(validator.Result("No such user exists.")));
+
+            if (currentUser.GetUserId() != Guid.Parse(id))
+            {
+                formatter.RenderJson(validator.Result("Sorry, you cannot edit another user's profile."), out string responseTxt);
+                httpHelper.Forbid(Response, responseTxt);
+                return null;
+            }
 
             validationResult = validator.IsBlank("username", username);
             if (validationResult.Key == false)
@@ -154,6 +168,7 @@ namespace SleekChat.Api.Controllers
             return Ok(formatter.Render(updatedUser, "User", Operation.Updated));
         }
 
+
         // DELETE: api/users/id
         [HttpDelete("{id}")]
         public ActionResult Delete([FromRoute] string id)
@@ -168,6 +183,13 @@ namespace SleekChat.Api.Controllers
 
             if (userData.GetUserById(Guid.Parse(id)) == null)
                 return NotFound(formatter.Render(validator.Result("No such user exists.")));
+
+            if (currentUser.GetUserId() != Guid.Parse(id))
+            {
+                formatter.RenderJson(validator.Result("Sorry, you cannot delete another user's profile."), out string responseTxt);
+                httpHelper.Forbid(Response, responseTxt);
+                return null;
+            }
 
             userData.DeleteUser(Guid.Parse(id));
             return Ok(formatter.Render(null, "User", Operation.Deleted));
